feat: implement core admin repository, RBAC services, and frontend scaffolding for property management and user settings

+++ Server/internal/api/handlers/saved_properties.go
@@ -35,7 +35,7 @@ func (h *Handler) AddSavedProperty(c *gin.Context) {
 
 	savedProperty := models.UserSavedProperty{
 		UserID:     user.ID,
-		PropertyID: uint(propertyID),
+		PropertyID: propertyID,
 	}
 
 	// db.Clauses(clause.OnConflict{DoNothing: true}).Create(&savedProperty)
